test(recovery): share permanent watchdog fixture (#728)
## Summary

- move the permanent isolated-watchdog recovery harness into the
existing shared recovery fixture
- remove duplicated lifecycle/task-executor setup from the
terminal-exhaustion spec
- preserve the distinct transient-recovery and terminal-exhaustion
assertions and cleanup behavior

## Root cause

Opcore normalizes nonblank source into five-line clone windows. Both
recovery specs began with the same `assert`, lifecycle, task-executor,
and fixture-destructuring setup, producing `clone-d5dcb9038dd73b67`.

## Validation

- focused recovery/ownership suite: 29 passing
- `npm run lint`: pass (155 warnings, 0 errors)
- `npm run typecheck`: pass
- `npm run test:unit`: 1685 passing, 18 pending
- `npm run protocol:check`: pass
- `npm run rust:check`: pass
- `opcore check --changed --base
05c50af0cd91a008889bd89498368b79f14d618f --json`: pass; no clone
diagnostic for either recovery spec

The tag-based `v6.7.0` comparison is reserved for the narrow release
candidate; on `dev` it also spans unrelated #717 Rust work.

Closes #727
